lib/ukvmem: Remove VMA implementation from headers
This change moves the definition of file and DMA VMA structs from the
public `<uk/vma_types.h>` header and into their respective
implementation source files.
